refactor: HitOrStand 패키지 view로 수정
