+++ src/main/java/blackjack/BlackJackApplication.java
@@ -4,10 +4,10 @@ import blackjack.model.game.BettingResult;
 import blackjack.model.game.BlackJackGame;
 import blackjack.model.game.Deck;
 import blackjack.model.game.DeckInitializer;
-import blackjack.model.game.HitOrStand;
 import blackjack.model.player.Dealer;
 import blackjack.model.player.Player;
 import blackjack.model.player.Players;
+import blackjack.view.HitOrStand;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import java.util.ArrayList;
